+++ src/onegov.election_day/onegov/election_day/formats/election/wabsti_proporz.py
@@ -427,9 +427,12 @@ def import_election_wabsti_proporz(
         return errors
 
     # Add the missing entities
-    remaining = entities.keys() - results.keys()
+    remaining = set(entities.keys())
+    if election.expats:
+        remaining.add(0)
+    remaining -= set(results.keys())
     for entity_id in remaining:
-        entity = entities[entity_id]
+        entity = entities.get(entity_id, {})
         district = entity.get('district', '')
         if election.domain == 'region':
             if not election.distinct: