+++ src/onegov.election_day/onegov/election_day/formats/election/wabstic_majorz.py
@@ -404,9 +404,12 @@ def import_election_wabstic_majorz(
 
     # Add the missing entities
     result_inserts = []
-    remaining = entities.keys() - added_results.keys()
+    remaining = set(entities.keys())
+    if election.expats:
+        remaining.add(0)
+    remaining -= set(added_results.keys())
     for entity_id in remaining:
-        entity = entities[entity_id]
+        entity = entities.get(entity_id, {})
         district = entity.get('district', '')
         if election.domain == 'region':
             if not election.distinct: